+++ service/relation.go
@@ -13,8 +13,8 @@ import (
 type FollowResponse struct {
 	ID            uint64 `json:"id"`
 	Name          string `json:"name"`
-	FollowCount   int    `json:"follow_count"`
-	FollowerCount int    `json:"follower_count"`
+	FollowCount   uint64 `json:"follow_count"`
+	FollowerCount uint64 `json:"follower_count"`
 	IsFollow      bool   `json:"is_follow"`
 }
